Refactor signup route
